requirements.txt
accelerate==0.19.0
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-socks==0.8.0
aiosignal==1.3.1
altair==4.2.2
ansi2html==1.8.0
anyio==3.6.2
appnope==0.1.3
argilla==1.4.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.1.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.0
cchardet==2.1.7
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
Chroma==0.2.0
chromadb==0.3.23
click==8.1.3
clickhouse-connect==0.5.24
comm==0.1.3
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.3.0
duckdb==0.8.0
elastic-transport==8.4.0
elasticsearch==8.7.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.7.3
fake-useragent==1.1.3
Faker==18.9.0
fastapi==0.94.1
favicon==0.7.0
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.3.2
fonttools==4.39.4
frozenlist==1.3.3
future==0.18.3
fuzzywuzzy==0.18.0
geographiclib==2.0
geojson==2.5.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.2.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.25.0
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-generativeai==0.1.0rc3
google-resumable-media==2.5.0
google-search-results==2.4.2
googleapis-common-protos==1.56.4
googletransx==2.4.2
greenlet==2.0.2
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.7.0
htbuilder==0.6.1
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.0
ipython==8.12.2
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kaleido==0.2.1
kiwisolver==1.4.4
langchain==0.0.125
libmagic==1.0
llama-index==0.6.0
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.1
markdownlit==0.0.7
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==9.0.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openai-whisper==20230124
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.0
plotly==5.14.1
pod-gpt==0.0.7
posthog==3.0.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycoingecko==3.1.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==10.0.1
Pympler==1.0.1
pyowm==3.3.0
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.9.0
PyPDF2==3.0.1
PyQt6==6.4.2
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.1
python-magic==0.4.27
python-pptx==0.6.21
python-socks==2.3.0
pytube==15.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
schedule==1.2.0
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.5
soundfile==0.11.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
st-annotated-text==4.0.0
stack-data==0.6.2
starlette==0.26.1
streamlit==1.22.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
stweet==2.1.1
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tor-python-easy==0.1.5
torch==2.0.1
torchvision==0.15.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
trio==0.22.0
trio-websocket==0.10.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
unstructured==0.5.4
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.0
uvloop==0.17.0
validators==0.20.0
vertexai==0.0.1
watchfiles==0.19.0
wcwidth==0.2.6
webdriver-manager==3.8.6
websockets==10.4
Werkzeug==2.3.4
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.0.9
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
